+++ lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:places/ui/screen/sight_details.dart';
 
-import 'package:places/ui/screen/sight_list_screen.dart';
+import 'package:places/ui/screens/sight_list_screen.dart';
 
 
 void main() {
