+++ lib/ui/res/app_assets.dart
@@ -3,4 +3,5 @@ abstract class AppAssets {
   static const calendar = 'lib/ui/res/icons/calendar.svg';
   static const favouriteDark = 'lib/ui/res/icons/heart_dark.svg';
   static const goIcon = 'lib/ui/res/icons/go.svg';
+  static const noImage = 'lib/ui/res/images/no_image.png';
 }
