+++ lib/ui/widgets/sight_details_image.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 import 'package:places/domain/sight.dart';
